% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/package.rokogiri.R, R/rokogiri.R
\docType{package}
\name{rokogiri}
\alias{rokogiri}
\alias{rokogiri-package}
\title{An XML generator DSL in R, inspired by Ruby's Nokogiri.}
\usage{
rokogiri(expr, output_type)
}
\arguments{
\item{expr}{expression. The R expression to compile to XML.}

\item{output_type}{character. Either \code{'list'} or \code{'xml'}.
By default, the latter.}

\item{enclos}{environment. The enclosing environment to use for evaluating
variables mentioned in the XML generation process. By default,
\code{parent.frame()}.}
}
\value{
a character string representing the compiled XML.
}
\description{
An XML generator DSL in R, inspired by Ruby's Nokogiri.

Compile an R expression into XML.
}
\examples{
xml <- rokogiri({
  note({
    to("Tove")
    from("Jani")
    heading("Reminder")
    body("Don't forget me this weekend!")
  })
})

expected_xml <- "
  <note>
    <to>Tove</to>
    <from>Jani</from>
    <heading>Reminder</heading>
    <body>Don't forget me this weekend!</body>
  </note>
"

remove_spacing <- function(x) { gsub("[ \\n]", "", x) }
stopifnot(remove_spacing(xml) == remove_spacing(expected_xml))
}

